+++ JsonToMidiConverter/Models/Song/Tempó.cs
@@ -2,7 +2,7 @@
 
 public sealed class Tempó : MeasureTempo
 {
-    public int Measure { get; set; }
+    public ushort Measure { get; set; }
     public float Position { get; set; }
     public bool Visible { get; set; }
     public bool Linear { get; set; }
